[feat] add main_test.cpp
