+++ test_shell/main_test.cpp
@@ -1,0 +1,24 @@
+#ifdef _DEBUG
+#include <cstdlib>
+#include "mock_ssd_adapter.h"
+#include "test_shell.h"
+
+using namespace testing;
+
+TEST(SampleTest, Addition) {
+    EXPECT_EQ(2 + 2, 4);
+}
+
+TEST(TS, ReadPass) {
+    MockSSDAdapter mockSSDAdapter;
+    TestShell testShell;
+
+    EXPECT_CALL(mockSSDAdapter, read(10))
+        .WillRepeatedly(Return("0xAAAABBBB"));
+}
+
+int main(int argc, char** argv) {
+    ::testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
+#endif
