+++ test_shell/test_shell.vcxproj
@@ -127,6 +127,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="main.cpp" />
+    <ClCompile Include="main_test.cpp" />
     <ClCompile Include="test_shell.cpp" />
   </ItemGroup>
   <ItemGroup>
